+++ src/org/mql/gc/dao/DonateurDao.java
@@ -8,10 +8,10 @@ import org.mql.gc.models.Donnateur;
 public interface DonateurDao {
 	public void create(Donnateur a) ;
 	public Donnateur update(Donnateur e) ;
-	public void delete(Long id) ;
+	public void delete(int id) ;
 	
 	public List<Donnateur> findAll() ;
-	public Donnateur findById(Long id) ;
+	public Donnateur findById(int id) ;
 	public Donnateur findByName(String name);
 	public Donnateur loginDonator(String email, String password);
 	public Donnateur findByEmail(String email);
